+++ src/python/rv/modules/metamodule.py
@@ -33,7 +33,7 @@ class UserDefined(Controller):
     def __init__(self, number):
         self.name = "user_defined_{}".format(number + 1)
         self.number = number + 6
-        super().__init__((0, 32768), 0, attached=False)
+        super().__init__((0, 44100), 0, attached=False)
 
     def attach(self, instance):
         self._attached = True
@@ -50,9 +50,8 @@ class UserDefinedProxy(Controller):
     def __get__(self, instance, owner):
         if instance is None:
             return self
-        else:
-            ctl = instance.user_defined[self.index]
-            return ctl.__get__(instance, owner)
+        ctl = instance.user_defined[self.index]
+        return ctl.__get__(instance, owner)
 
     def __set__(self, instance, value):
         if instance is not None:
@@ -113,14 +112,17 @@ class MetaModule(BaseMetaModule, Module):
         def update_user_defined_controllers(metamodule):
             project = metamodule.project
             items = zip(metamodule.mappings.values, metamodule.user_defined)
-            for mapping, user_defined_controller in items:
+            for idx, (mapping, user_defined_controller) in enumerate(items):
+                if idx == metamodule.user_defined_controllers:
+                    break
                 if mapping.module == 0 or mapping.module >= len(project.modules):
                     continue
                 mod = project.modules[mapping.module]
                 if not mod:
                     continue
-                controller_index = mapping.controller - 1
-                controller = list(mod.controllers.values())[controller_index]
+                controller_index = mapping.controller
+                controller_values = list(mod.controllers.values())
+                controller = controller_values[controller_index]
                 user_defined_controller.value_type = controller.instance_value_type(mod)
                 user_defined_controller.default = controller.default
                 metamodule.controller_values[
@@ -164,7 +166,7 @@ class MetaModule(BaseMetaModule, Module):
         ]
         super(MetaModule, self).__init__(**kwargs)
         self.mappings = self.MappingArray()
-        self.project = project if project else Project()
+        self.project = project or Project()
         self.project.metamodule = self
 
     def __getattr__(self, key):